Added Search Functionality in Canteen page
+ Refactored Project file structure
+ Added Error handling
+ Removed unnecessary methods
