+++ lib/main.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'canteen.dart';
-import 'upcoming.dart';
-import 'settings.dart';
-import 'about.dart';
-import 'color_schemes.dart';
+import 'package:vit_mess/screens/home_screen.dart';
+
+import 'utils/color_schemes.dart';
 
 void main() {
   runApp(const MyApp());
@@ -18,85 +15,12 @@ class MyApp extends StatefulWidget {
 }
 
 class MyAppState extends State<MyApp> {
-  void readselectedMess() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    int? savedselectedMess = prefs.getInt("selectedMess");
-    if (savedselectedMess != null) {
-      setState(() {
-        selectedMess = savedselectedMess;
-      });
-    }
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    readselectedMess();
-  }
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      theme: ThemeData(useMaterial3: true, colorScheme: lightColorScheme),
-      darkTheme: ThemeData(useMaterial3: true, colorScheme: darkColorScheme),
-      debugShowCheckedModeBanner: false,
-      home: DefaultTabController(
-        length: 2,
-        child: Scaffold(
-          appBar: AppBar(
-            bottom: const TabBar(
-              tabs: [
-                Tab(text: "Upcoming", icon: Icon(Icons.restaurant)),
-                Tab(text: "Canteen", icon: Icon(Icons.storefront)),
-              ],
-            ),
-            centerTitle: true,
-            title: const Text('VIT Bhopal Mess',
-                style: TextStyle(fontSize: 25, fontWeight: FontWeight.w400)),
-            actions: [
-              Builder(builder: (BuildContext context) {
-                return PopupMenuButton(
-                  itemBuilder: (BuildContext context) {
-                    return [
-                      const PopupMenuItem(
-                        value: 'Settings',
-                        child: Text(
-                          'Settings',
-                          style: TextStyle(fontSize: 17),
-                        ),
-                      ),
-                      const PopupMenuItem(
-                        value: 'About',
-                        child: Text(
-                          'About',
-                          style: TextStyle(fontSize: 17),
-                        ),
-                      ),
-                    ];
-                  },
-                  onSelected: (value) {
-                    if (value == 'Settings') {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => const Settings()),
-                      );
-                    } else if (value == 'About') {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => const About()),
-                      );
-                    }
-                  },
-                );
-              }),
-            ],
-          ),
-          body: const TabBarView(
-            children: [Upcoming(), Canteen()],
-          ),
-        ),
-      ),
-    );
+        theme: ThemeData(useMaterial3: true, colorScheme: lightColorScheme),
+        darkTheme: ThemeData(useMaterial3: true, colorScheme: darkColorScheme),
+        debugShowCheckedModeBanner: false,
+        home: const HomeScreen());
   }
 }
